feat: add loading overlay and service
This commit introduces a new `LoadingOverlay` widget and a `BaseLoadingService`.

The `LoadingOverlay` widget can be used to display a loading indicator on top of its child. It's intended to be used with `MaterialApp.builder`.

The `BaseLoadingService` provides a default implementation for managing loading states and displaying a loading widget. It includes a `ButterflyLoadingController` to manage the loading state and allows customization of the loading widget.

Additionally, the `BaseCoreService` has been updated:
- It now has a static `instance` getter for easy access.
- It provides default implementations for `errorHandlerService` (using `DefaultErrorHandler`) and a new `loadingService` (using `DefaultLoadingService`).