feat: add example for go_router package
This commit introduces an example application demonstrating the usage of the `go_router` package for navigation in a Flutter application.

The example includes:
- `HomeScreen` and `SecondScreen` as placeholder screens.
- A `RouteService` class (currently empty) intended for route management.

Dependencies in `pubspec.yaml` have been updated to the latest versions of `go_router`, `flutter_lints`, and `go_router_builder`. The local path dependency for `core_management` has been removed.