Refactor: Use `InputDecorationThemeData` for input decoration theme
The `InputDecorationTheme` class in Flutter defines the default appearance of `InputDecorator`s.

This commit updates the `base_theme.dart` file to use `InputDecorationThemeData` instead of `InputDecorationTheme` for the `inputDecorationTheme` property. This change aligns with the Flutter framework's recommendation to use `InputDecorationThemeData` for describing the actual configuration of an input decoration theme.

The `appBarTheme` property has also been updated to use `AppBarThemeData` for consistency.